Design: 콜라보레이터 팝오버 모달로 수정
