+++ src/app/list/[listId]/_components/ListDetailOuter/ListInformation.css.ts
@@ -90,4 +90,5 @@ export const infoDetailWrapper = style({
 
 export const collaboratorWrapper = style({
   display: 'flex',
+  cursor: 'pointer',
 });
